rename parser.py file so it does not override the built-in python parser
