+++ msclam_parser.py
